lc0258: Add Digits
