+++ src/main/java/com/zea7ot/lc/lvl1/lc0258/SolutionApproach0Math.java
@@ -1,9 +1,12 @@
 /**
  * https://leetcode.com/problems/add-digits/
+ * 
+ * Time Complexity:     O(1)
+ * Space Complexity:    O(1)
  */
 package com.zea7ot.lc.lvl1.lc0258;
 
-public class SolutionApproachMath {
+public class SolutionApproach0Math {
     public int addDigits(int num) {
         return 1 + (num - 1) % 9;
     }
