+++ src/main/java/com/zea7ot/lc/tags/algorithm/math/summary.txt
@@ -2,6 +2,8 @@
 
 ### Math
 
+1   0258    https://leetcode.com/problems/add-digits/
+
 2   0441    https://leetcode.com/problems/arranging-coins/
 2   0463    https://leetcode.com/problems/island-perimeter/
 2   1344    https://leetcode.com/problems/angle-between-hands-of-a-clock/
