+++ src/main/java/com/zea7ot/lc/tags/algorithm/search/depth_first_search/summary.txt
@@ -4,6 +4,7 @@ References:
 
 ### Depth First Search
 
+1   0258    https://leetcode.com/problems/add-digits/
 1   0559    https://leetcode.com/problems/maximum-depth-of-n-ary-tree/
 1   0671    https://leetcode.com/problems/second-minimum-node-in-a-binary-tree/
 1   0700    https://leetcode.com/problems/search-in-a-binary-search-tree/
